fix sync🤷‍

+++ src/init.cpp
@@ -1645,7 +1645,6 @@ bool AppInitMain(NodeContext& node, interfaces::BlockAndHeaderTipInfo* tip_info)
 
         do {
             const int64_t load_block_index_start_time = GetTimeMillis();
-            dev::eth::ChainParams cp(chainparams.EVMGenesisInfo());
             try {
                 LOCK(cs_main);
                 chainman.Reset();
@@ -1705,50 +1704,58 @@ bool AppInitMain(NodeContext& node, interfaces::BlockAndHeaderTipInfo* tip_info)
                     break;
                 }
 
+
+
                 /////////////////////////////////////////////////////////// qtum
-                if((args.IsArgSet("-dgpstorage") && args.IsArgSet("-dgpevm")) || (!args.IsArgSet("-dgpstorage") && args.IsArgSet("-dgpevm")) ||
-                  (!args.IsArgSet("-dgpstorage") && !args.IsArgSet("-dgpevm"))){
-                    fGettingValuesDGP = true;
-                } else {
-                    fGettingValuesDGP = false;
+                LOCK(cs_main);
+                CChain& active_chain = chainman.ActiveChain();
+                if (active_chain.Height() > chainparams.GetConsensus().nSmartActivationBlock){
+                    if((args.IsArgSet("-dgpstorage") && args.IsArgSet("-dgpevm")) || (!args.IsArgSet("-dgpstorage") && args.IsArgSet("-dgpevm")) ||
+                      (!args.IsArgSet("-dgpstorage") && !args.IsArgSet("-dgpevm"))){
+                        fGettingValuesDGP = true;
+                    } else {
+                        fGettingValuesDGP = false;
+                    }
+
+                    dev::eth::NoProof::init();
+                    fs::path qtumStateDir = gArgs.GetDataDirNet() / "stateQtum";
+                    bool fStatus = fs::exists(qtumStateDir);
+                    const std::string dirQtum(qtumStateDir.string());
+                    const dev::h256 hashDB(dev::sha3(dev::rlp("")));
+                    dev::eth::BaseState existsQtumstate = fStatus ? dev::eth::BaseState::PreExisting : dev::eth::BaseState::Empty;
+                    globalState = std::unique_ptr<QtumState>(new QtumState(dev::u256(0), QtumState::openDB(dirQtum, hashDB, dev::WithExisting::Trust), dirQtum, existsQtumstate));
+
+                    dev::eth::ChainParams cp(chainparams.EVMGenesisInfo());
+                    globalSealEngine = std::unique_ptr<dev::eth::SealEngineFace>(cp.createSealEngine());
+
+                    pstorageresult.reset(new StorageResults(qtumStateDir.string()));
+                    if (fReset) {
+                        pstorageresult->wipeResults();
+                    }
+
+                    fRecordLogOpcodes = args.IsArgSet("-record-log-opcodes");
+                    fIsVMlogFile = fs::exists(gArgs.GetDataDirNet() / "vmExecLogs.json");
+
+                    if (fAddressIndex != args.GetBoolArg("-addrindex", DEFAULT_ADDRINDEX)) {
+                        strLoadError = _("You need to rebuild the database using -reindex to change -addrindex");
+                        break;
+                    }
+
+                    // Check for changed -logevents state
+                    if (fLogEvents != args.GetBoolArg("-logevents", DEFAULT_LOGEVENTS) && !fLogEvents) {
+                        strLoadError = _("You need to rebuild the database using -reindex to enable -logevents");
+                        break;
+                    }
+
+                    if (!args.GetBoolArg("-logevents", DEFAULT_LOGEVENTS))
+                    {
+                        pstorageresult->wipeResults();
+                        pblocktree->WipeHeightIndex();
+                        fLogEvents = false;
+                        pblocktree->WriteFlag("logevents", fLogEvents);
+                    }
+                    ///////////////////////////////////////////////////////////////
                 }
-
-                dev::eth::NoProof::init();
-                fs::path qtumStateDir = gArgs.GetDataDirNet() / "stateQtum";
-                bool fStatus = fs::exists(qtumStateDir);
-                const std::string dirQtum(qtumStateDir.string());
-                const dev::h256 hashDB(dev::sha3(dev::rlp("")));
-                dev::eth::BaseState existsQtumstate = fStatus ? dev::eth::BaseState::PreExisting : dev::eth::BaseState::Empty;
-                globalState = std::unique_ptr<QtumState>(new QtumState(dev::u256(0), QtumState::openDB(dirQtum, hashDB, dev::WithExisting::Trust), dirQtum, existsQtumstate));
-                globalSealEngine = std::unique_ptr<dev::eth::SealEngineFace>(cp.createSealEngine());
-
-                pstorageresult.reset(new StorageResults(qtumStateDir.string()));
-                if (fReset) {
-                    pstorageresult->wipeResults();
-                }
-
-                fRecordLogOpcodes = args.IsArgSet("-record-log-opcodes");
-                fIsVMlogFile = fs::exists(gArgs.GetDataDirNet() / "vmExecLogs.json");
-
-                if (fAddressIndex != args.GetBoolArg("-addrindex", DEFAULT_ADDRINDEX)) {
-                    strLoadError = _("You need to rebuild the database using -reindex to change -addrindex");
-                    break;
-                }
-
-                // Check for changed -logevents state
-                if (fLogEvents != args.GetBoolArg("-logevents", DEFAULT_LOGEVENTS) && !fLogEvents) {
-                    strLoadError = _("You need to rebuild the database using -reindex to enable -logevents");
-                    break;
-                }
-
-                if (!args.GetBoolArg("-logevents", DEFAULT_LOGEVENTS))
-                {
-                    pstorageresult->wipeResults();
-                    pblocktree->WipeHeightIndex();
-                    fLogEvents = false;
-                    pblocktree->WriteFlag("logevents", fLogEvents);
-                }
-                ///////////////////////////////////////////////////////////////
 
                 // At this point we're either in reindex or we've loaded a useful
                 // block tree into BlockIndex()!
@@ -1811,6 +1818,7 @@ bool AppInitMain(NodeContext& node, interfaces::BlockAndHeaderTipInfo* tip_info)
             CChain& active_chain = chainman.ActiveChain();
             if (active_chain.Height() > chainparams.GetConsensus().nSmartActivationBlock)
             {
+                dev::eth::ChainParams cp(chainparams.EVMGenesisInfo());
                 if(active_chain.Tip() != nullptr){
                     globalState->setRoot(uintToh256(active_chain.Tip()->hashStateRoot));
                     globalState->setRootUTXO(uintToh256(active_chain.Tip()->hashUTXORoot));
